Using a more precise capturing group for `T_IDENTIFIER`
A `T_IDENTIFIER` represents:

 * a constant
 * a function name
 * a function name prefixed with `\`
 * a class name
 * a class name prefixed with `\`
 * a class constant
 * a class constant, with class name prefixed with `\`

The current capturing regular expression is too simplistic, and it
leads to edge cases where `Foo\42.5` are captured as separate tokens.

In this patch, we deprecate `DocLexer::T_NAMESPACE_SEPARATOR` (which
shouldn't be looked up directly anymore), and tell the parser to work
with wider `T_IDENTIFIER` tokens instead.

Note that a test is also being removed. That is intentional, since the
test verifies a parser issue when parsing `@Foo\3.42`:

 * previously, the tokens would be `[@ Foo\3.42]`
 * now they are `[@, Foo, \, 3.42]`

This also means that the above will be parsed as `@Foo`.