Remove calls to TestCase::at()
It has been deprecated as of PHPUnit 9

Note that we still rely on the call order here, I am not the original
author of this test but the order looks like it is pretty important for
this test, or maybe at least for understanding it.
Also note that we no longer check that save() is called after all calls
to fetch().