Add compatibility with named parameters and constructor property promotion
With PHP 8 the combination of named parameters and constructor property
promotion will be very cool for the new Attributes feature, but also for
Doctrine's Annotation feature. Especially if classes are to be used
for both Attributes and Annotations, then the DocParser needs
a compatible support for named parameters in annotation class constructors.

For backwards compatibility this is done with a marker interface that
uses a different instantiation logic than the current one. Currently
all annotations properties are passed as a single array argument
$values.

This patch uses reflection to match annotation values to class
constructor parameters. For PHP 8 the integration of named arguments
with spread operator is used.

A PHP 8 class that could be read using Doctrine Annotations would
become:

/** @Annotation */
class Table implements NamedArgumentConstructorAnnotation
{
    public function __construct(
        public string $name,
        public array $indexes = []
    ) {}
}
/** @Table(name="users") */
/** @Table(name="users", indexes={"foo"}) */