Handle positional argument following named ones
Use `array_pop($positional_arguments)` instead of
`$positional_arguments[0]` to access the only element from the
positional arguments array.
This fixes a bug that arises when there is one positional argument after
named ones, example: `@Name(foo=“bar”, “foo”)`, the positional arguments
array is indexed by position. So in this case, the index is `1` and not
`0`, therefore, trying to access the value with `[0]` doesn't work.

Fixes #405