Drop Lexer 1
This is not a big deal because doctrine/lexer 1 has the same version
constraints as doctrine/lexer 2.
Note that I'm removing the BC-layer for DocLexer::peek() and DocLexer::glimpse():
in retrospect, I think it was a bad idea because there was already a
BC-break with DocLexer::$lookahead and DocLexer::$token.